Refactor structure for maintainability
Refactor, add cachelib
