+++ docs/api.rst
@@ -9,4 +9,11 @@ Anything documented here is part of the public API that Flask-Session provides, 
 .. autoclass:: Session
    :members: init_app
 
-.. autoclass:: flask_session.sessions.ServerSideSession
+.. autoclass:: flask_session.base.ServerSideSession
+
+.. autoclass:: flask_session.redis.RedisSessionInterface
+.. autoclass:: flask_session.memcached.MemcachedSessionInterface
+.. autoclass:: flask_session.filesystem.FileSystemSessionInterface
+.. autoclass:: flask_session.cachelib.CacheLibSessionInterface
+.. autoclass:: flask_session.mongodb.MongoDBSessionInterface
+.. autoclass:: flask_session.sqlalchemy.SqlAlchemySessionInterface
