+++ docs/config_exceptions.rst
@@ -1,12 +1,7 @@
-Storage exceptions
-===================
-
-For various reasons, database operations can fail. When a database operation fails, the database client will raise an Exception.
-
 Retries
 --------
 
-Upon an Exception, Flask-Session will retry with backoff up to 3 times for SQL based storage. If the operation still fails after 3 retries, the Exception will be raised.
+Only for SQL based storage, upon an exception, Flask-Session will retry with backoff up to 3 times. If the operation still fails after 3 retries, the exception will be raised.
 
 For other storage types, the retry logic is either included or can be configured in the client setup. Refer to the client's documentation for more information.
 
@@ -22,6 +17,7 @@ Redis example with retries on certain errors:
         ConnectionError,
         TimeoutError
     )
+    ...
 
     retry = Retry(ExponentialBackoff(), 3)
     SESSION_REDIS = Redis(host='localhost', port=6379, retry=retry, retry_on_error=[BusyLoadingError, ConnectionError, TimeoutError])
