+++ src/flask_session/filesystem/__init__.py
@@ -1,0 +1,1 @@
+from .filesystem import FileSystemSessionInterface, FileSystemSession
