+++ src/flask_session/memcached/__init__.py
@@ -1,0 +1,1 @@
+from .memcached import MemcachedSessionInterface, MemcachedSession
