+++ src/flask_session/mongodb/__init__.py
@@ -1,0 +1,1 @@
+from .mongodb import MongoDBSessionInterface, MongoDBSession
