+++ src/flask_session/redis/__init__.py
@@ -1,0 +1,1 @@
+from .redis import RedisSessionInterface, RedisSession
