+++ src/flask_session/sqlalchemy/__init__.py
@@ -1,0 +1,1 @@
+from .sqlalchemy import SqlAlchemySessionInterface, SqlAlchemySession
