+++ tests/test_basic.py
@@ -3,13 +3,6 @@ import flask_session
 import pytest
 
 
-def test_tot_seconds_func():
-    import datetime
-
-    td = datetime.timedelta(days=1)
-    assert flask_session.sessions.total_seconds(td) == 86400
-
-
 def test_null_session():
     """Invalid session should fail to get/set the flask session"""
     with pytest.raises(ValueError):
