+++ tests/test_cachelib.py
@@ -1,24 +1,24 @@
 import tempfile
 
 import flask
-from flask_session.filesystem import FileSystemSession
+from flask_session.cachelib import CacheLibSession
 
 
-class TestFileSystemSession:
+class TestCachelibSession:
 
     def retrieve_stored_session(self, key, app):
         return app.session_interface.cache.get(key)
 
     def test_filesystem_default(self, app_utils):
         app = app_utils.create_app(
-            {"SESSION_TYPE": "filesystem", "SESSION_FILE_DIR": tempfile.gettempdir()}
+            {"SESSION_TYPE": "cachelib", "SESSION_SERIALIZATION_FORMAT": "json"}
         )
 
         # Should be using FileSystem
         with app.test_request_context():
             assert isinstance(
                 flask.session,
-                FileSystemSession,
+                CacheLibSession,
             )
             app_utils.test_session(app)
